pyproject.toml
[project]
name = "ocr-agentic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Ming Loon Lee",email = "leemingloon93@gmail.com"}
]
package-mode = false   # <-- Add this line
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "computer (>=1.0.0,<2.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "deep (>=0.10,<0.11)",
    "learning (==0.0.*)",
    "onnxruntime (>=1.24.1,<2.0.0)",
    "paddleocr (>=3.4.0,<4.0.0)",
    "paddlepaddle (>=3.3.0,<4.0.0)",
    "nlp (>=0.4.0,<0.5.0)",
    "embeddings (>=0.0.8,<0.0.9)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "agentic (>=0.0.2,<0.0.3)",
    "ai (>=1.2,<2.0)",
    "anthropic (>=0.79.0,<0.80.0)",
    "evaluation (>=0.0.2,<0.0.3)",
    "datasets (>=4.5.0,<5.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "data (>=0.4,<0.5)",
    "pydantic (>=2.12.5,<3.0.0)",
    "monitoring (>=0.9.12,<0.10.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "opentelemetry-api (>=1.39.1,<2.0.0)",
    "opentelemetry-sdk (>=1.39.1,<2.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "multimodal (>=0.0.13,<0.0.14)",
    "retrieval (>=0.0.1.dev6,<0.0.2)",
    "credit (>=0.0.1,<0.0.2)",
    "risk (>=0.1.4,<0.2.0)",
    "xgboost (>=3.2.0,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "prometheus-client (>=0.24.1,<0.25.0)",
    "boto3 (>=1.42.49,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "optional (>=0.0.1,<0.0.2)",
    "jupyter (>=1.1.1,<2.0.0)",
    "notebook (>=7.5.3,<8.0.0)",
    "llama-index-core (>=0.14.14,<0.15.0)",  # Should be added here
    "faiss-cpu",
    "llama-index (>=0.14.14,<0.15.0)",
    "evidently (>=0.7.20,<0.8.0)",
    "ragas (>=0.4.3,<0.5.0)",
    "sagemaker (>=3.4.1,<4.0.0)",
    "transformers (>=4.41.0,<5.0.0)",
    "sentence-transformers (<5.0.0)",
    "flagembedding (>=1.3.5,<2.0.0)",
    "langchain (<0.1.0)",
    "langchain-anthropic (>=1.3.3,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "ocr_agentic_rag", from = "src" }]
